+++ .gitignore
@@ -1,3 +1,5 @@
-__pycache__/
+*__pycache__/
 **/*.npz
 old_tests/
+*.egg-info/
+build/
