--- onnx/pscan.py
@@ -1,1 +1,0 @@
-../pscan.py
