+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='mambapy',
-    version='0.1.2',
+    version='1.0.0',
     packages=find_packages(),
     install_requires=[
         'torch',
